+++ internal/commands/make_entity.go
@@ -11,16 +11,17 @@ import (
 
 func NewMakeEntityCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "make entity <Name>",
+		Use:   "entity <Name>",
 		Short: "Generates a domain entity",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			name := args[0]
+			namePascal := internal.ToPascalCase(name)
 
 			out := filepath.Join("internal/domain/entities", internal.ToSnake(name)+".go")
 
 			data := map[string]any{
-				"Name": name,
+				"Name": namePascal,
 			}
 
 			if err := internal.WriteTemplate(templates.FS, "entity/entity.go.tpl", out, data); err != nil {